split dftx_balance into balance and account
