+++ packages/jellyfish-transaction/__tests__/script/defi/dftx_balance/UtxosToAccount.test.ts
@@ -1,5 +1,5 @@
 import { SmartBuffer } from 'smart-buffer'
-import { CUtxosToAccount, UtxosToAccount } from '../../../../src/script/defi/dftx_balance'
+import { CUtxosToAccount, UtxosToAccount } from '../../../../src/script/defi/dftx_account'
 import { OP_CODES, toBuffer, toOPCodes } from '../../../../src/script'
 import BigNumber from 'bignumber.js'
 import { OP_DEFI_TX } from '../../../../src/script/defi'
